+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.1.1] - 2025-06-12
+
+- Added support for the correct image file paths for images
+- Added Jekyll script
+
 ## [0.1.0] - 2025-06-10
 
 - Initial release